Added code generation for assignment statements with a substring operation on the lhs.  e.g.   blah(4:6) = 'abc'
the changes are primarily in codegen.c, but I also made changes
to f2jparse.y and typecheck.c.

I added a 'Character' data type so that I can represent the
type returned by the intrinsic function 'CHAR'.  This helps in
code generation so that we dont try to generate something
like "somechar.charAt(0)", which may happen if we assume that
every character type is a string.