-changed the parsing of cilist (for READ/WRITE), olist (for OPEN),  and cllist (for CLOSE) to be more flexible (and correct)
-finished implementing ERR, END, and IOSPEC specifiers for READ,
 WRITE, OPEN, and CLOSE statements.  this involved a lot of hacking
 on the back end and hasn't been thoroughly tested, so it could
 be broken for some cases.

-implementing the changes above required changing the Util API
 a bit, so take note of that if you happen to be calling stuff
 from user code (not likely).  also if you have old f2j-generated
 code, then you might need to recompile when installing the new
 f2jutil.jar runtime library (depending on your i/o usage).

-removed some obsolete code

in terms of I/O, the main item left to be implemented is
direct access file I/O (and the corresponding REC and RECL
specifiers).  then everything needs to be tested.  it would
also be a good idea to make sure LAPACK still translates
correctly after all these changes.

keith